requirements.txt
uvicorn==0.29.0
fastapi==0.110.1
pydantic==2.7.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29